requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
addict==2.2.1
click==7.0
jinja2==2.10.3
markupsafe==1.1.1         # via jinja2
pyyaml==5.1.2
